% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllEvents_RNASeq_MP.R
\name{AllEvents_RNASeq_MP}
\alias{AllEvents_RNASeq_MP}
\title{Alternative splicing multi-path events detected by EventPointer}
\format{A \code{list} object \code{AllEvents_RNASeq[[i]][[j]]} displays the jth splicing event for the ith gene.}
\usage{
data(AllEvents_RNASeq_MP)
}
\value{
AllEvents_RNASeq_MP object contains all the detected alternative splicing events using EventPointer
methodology for multi-path events. The splicing events where detected using the BAM files from the dataset published in Seshagiri et al. 2012 and
used in the \code{SGSeq} R package vignette.
}
\description{
Alternative splicing multi-path events detected by EventPointer
}
