% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateEvents.R
\name{annotateEvents}
\alias{annotateEvents}
\title{Internal function called by EvenPointer}
\usage{
annotateEvents(Events, PSR_Gene, Junc_Gene, Gxx)
}
\arguments{
\item{Events}{Detected alternative splicing events for a particular gene}

\item{PSR_Gene}{Exon probes mapped to the gene that is being analyzes}

\item{Junc_Gene}{Junction probes mapped to the gene that is being analyzes}

\item{Gxx}{Identifier for the gene}
}
\value{
Returns a list with the information required for the .flat file and EventsFound.txt
}
\description{
Function to obtain all the information for the alternative splicing events for a particular gene
}
\keyword{internal}

