% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateEvents_RNASeq.R
\name{AnnotateEvents_RNASeq}
\alias{AnnotateEvents_RNASeq}
\title{Internal function called by EvenPointer}
\usage{
AnnotateEvents_RNASeq(Events)
}
\arguments{
\item{Events}{Detected alternative splicing events for a particular gene}
}
\value{
list with the information required for the EventsFound.txt file
}
\description{
Function to obtain all the information for the alternative splicing events for a particular gene for sequencing data
}
\keyword{internal}

