% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayDatamultipath.R
\name{ArrayDatamultipath}
\alias{ArrayDatamultipath}
\title{Preprocessed arrays data with multi-path events}
\format{A \code{data.frame} with preprocessed arrays data. The preprocessing
was done using \code{aroma.affymetrix}. See the package vignette for the preprocessing
pipeline}
\usage{
data(ArrayDatamultipath)
}
\value{
ArrayDatamultipath object contains preprocessed junction arrays data. The preprocessing was
done using \code{aroma.affymetrix} R package, refer to EventPointer vignette for the pipeline
used for the preprocessing. The data corresponds to 4 samples from the SUM149 Cell line
hybridized to the HTA 2.0 Affymetrix array. The first two samples are control and the
second ones are treated.
}
\description{
Preprocessed arrays data with multi-path events
}
