% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraysData.R
\name{ArraysData}
\alias{ArraysData}
\title{Preprocessed arrays data}
\format{
A \code{data.frame} with preprocessed arrays data.
The preprocessing was done using \code{aroma.affymetrix}.
See the package vignette for the preprocessing pipeline.
}
\usage{
data(ArraysData)
}
\value{
ArraysData object contains preprocessed junction arrays data.
The preprocessing was done using \code{aroma.affymetrix} R package,
refer to EventPointer vignette for the pipeline used for the preprocessing.
The data corresponds to 4 samples from the SUM149 Cell line hybridized to
the HTA 2.0 Affymetrix array. The first two samples are control and
the second ones are treated.
}
\description{
Preprocessed arrays data
}
