% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyEvents.R
\name{ClassifyEvents}
\alias{ClassifyEvents}
\title{Internal function called by EvenPointer}
\usage{
ClassifyEvents(SG, Events)
}
\arguments{
\item{Events}{Detected alternative splicing events for a particular gene}

\item{PSR_Gene}{Exon probes mapped to the gene that is being analyzes}

\item{Junc_Gene}{Junction probes mapped to the gene that is being analyzes}

\item{Gxx}{Identifier for the gene}
}
\value{
Events list with Type slot
}
\description{
Function to classify the alternative splicing events into the canonical categories: cassette exon, alternative 3', alternative 5',
alternative start site, alternative termination site, retained intron, mutually exclusive exons or complex events
}
\keyword{internal}

