% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomainEnrichment.R
\name{DomainEnrichment}
\alias{DomainEnrichment}
\title{Protein Domains Enrichment Analysis}
\usage{
DomainEnrichment(Events)
}
\arguments{
\item{Events}{The data.frame obtained with the EventPointer function}
}
\value{
Result List with 4 elements.

UpRegulated_Domain: (data.frame) Top 20 enriched domains with a Z value > 0

DownRegulated_Domain: (data.frame) Top 20 enriched domains with a Z value < 0

Pvalues: (vector) with the pvalues for all the domains

Zvalues: (vector) with the zvalues for all the domains
}
\description{
DomainEnrichment performs the analysis to identify the protein domains affected by the alternative splicing events returned by EventPointer function.
The enrichment analysis is done by using a Wilcoxon Rank Sum Test
}
\examples{
\dontrun{
DomainEnrichment(Events)}
}

