% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAbsoluteConc.R
\name{estimateAbsoluteConc}
\alias{estimateAbsoluteConc}
\title{Internal function called by EvenPointer}
\usage{
estimateAbsoluteConc(Signal1, Signal2, SignalR, lambda)
}
\arguments{
\item{Signal1}{Signal for Path 1}

\item{Signal2}{Signal for Path 2}

\item{SignalR}{Signal for Path R}

\item{lambda}{Penalty factor}
}
\value{
list with concentrations of both isoforms and relative error of the estimation
}
\description{
Estimate the concentration of the isoforms mapped to each of the three paths given the corresponding signal.
}
\keyword{internal}

