% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventPointerBAM_IGV.R
\name{EventPointerBAM_IGV}
\alias{EventPointerBAM_IGV}
\title{EventPointerBAM IGV Visualization}
\usage{
EventPointerBAM_IGV(SG_RNASeq, EventsTxt, PathGTF)
}
\arguments{

\item{SG_RNASeq}{SgFC valriable from EventsDetection_BAM() function. Contains splicing graphs components.}

\item{EventsTxt}{Path to TotalEventsFound.txt file generated with EventDetection function}

\item{PathGTF}{Directory where to write the GTF file.}
}
\value{
One .gtf file is written to the specified directory in PathGTF. The created file: 1) paths_RNASeq.gtf : GTF file representing the alternative splicing events.
}
\description{
Generates of files to be loaded in IGV for visualization and interpretation of events
}
\examples{
\dontrun{
  data(SgFC)
   # IGV Visualization

   EventsTxt<-paste(system.file('extdata',package='EventPointer'),'/TotalEventsFound_RNASeq.txt',sep='')
   PathGTF<-tempdir()
   EventPointerBAM_IGV(Events,SG_RNASeq,EventsTxt,PathGTF)
   }
}
