% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsGTFfromTrancriptomeGTF.R
\name{EventsGTFfromTrancriptomeGTF}
\alias{EventsGTFfromTrancriptomeGTF}
\title{Events .gtf from trancriptome .gtf}
\usage{
EventsGTFfromTrancriptomeGTF(inputFile = NULL, Transcriptome = NULL,
  Pathtxt = NULL, PathGTF = NULL)
}
\arguments{
\item{inputFile}{If input is GTF, inputFile should point to the GTF file to be used.}

\item{Transcriptome}{the name of the transcriptome}

\item{Pathtxt}{Directory to save the .txt of the events founded}

\item{PathGTF}{Directory where the output will be saved}
}
\value{
a list containing four elements: three sparce matrices that relate which isoforms build up the paths 
(path1,path2 and pathRef) of each event. The fourth element contains the name of the reference annotation: 
only appear the name of the transcript.
}
\description{
Events .gtf from trancriptome .gtf
}
\examples{
   PathFiles<-system.file("extdata",package="EventPointer")
   inputFile <- paste(PathFiles,"/gencode.v24.ann_2genes.gtf",sep="")
   Transcriptome <- "Gencode24_2genes"
   Pathtxt <- tempdir()
   PathGTF <- tempdir()
   
   # Run the function 
   
   EventXtrans <- EventsGTFfromTrancriptomeGTF(inputFile = inputFile,
                                               Transcriptome = Transcriptome,
                                               Pathtxt=Pathtxt,PathGTF=PathGTF)


}
