% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventXtrans.R
\name{EventXtrans}
\alias{EventXtrans}
\title{Relationship between isoforms and events.}
\format{
A \code{list} object \code{EventXtrans[[1]]} displays the isoform
that build up the path1 of each event.
}
\usage{
data(EventXtrans)
}
\value{
EventXtrans object contains the relationship between the
isoforms and the events. It is a list of 4 elements. the first three
stored sparse matrices relating the isoforms with the events. The fourth
element stores de names of the reference annotation used (isoforms names).
}
\description{
Relationship between isoforms and events.
}
