% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTriplets.R
\name{findTriplets}
\alias{findTriplets}
\title{Internal function called by EvenPointer}
\usage{
findTriplets(randSol, tol = 1e-08)
}
\arguments{
\item{randSol}{Flux traversing the edges of the splicing graph}

\item{tol}{Tolerance used}
}
\value{
list with all the triplets detected for a particular gene
}
\description{
Function to identify the triplets that represent alternative splicing events.
}
\keyword{internal}

