% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2Cdf.r
\name{flat2Cdf}
\alias{flat2Cdf}
\title{Internal function called by EvenPointer}
\usage{
flat2Cdf(file, chipType, tags = NULL, rows = 2560, cols = 2560,
  verbose = 10, xynames = c("X", "Y"), gcol = 5, ucol = 6, splitn = 4,
  col.class = c("integer", "character")[c(1, 1, 1, 2, 2, 2)], ...)
}
\arguments{
\item{file}{.flat file}

\item{chipType}{name of the array}

\item{tags}{tags to identify cdf file}

\item{rows}{rows in the array}

\item{cols}{columns in the array}

\item{verbose}{output to user}

\item{xynames}{identifiers for chromosome X and Y}

\item{gcol}{Internal parameter}

\item{ucol}{Internal parameter}

\item{splitn}{Internal parameter}

\item{col.class}{Internal parameter}
}
\value{
CDF file
}
\description{
Function to create the CDF file given the corresponding .flat file
}
\keyword{internal}

