% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbootstrapdata.R
\name{getbootstrapdata}
\alias{getbootstrapdata}
\title{getbootstrapdata}
\usage{
getbootstrapdata(PathSamples, type)
}
\arguments{
\item{PathSamples}{A vector with the complete directory to the folder of the output of kallisto/salmon.}

\item{type}{'kallisto' or 'salmon'.}
}
\value{
A list containing the quantification data with the bootstrap information.
}
\description{
Function to load the values of the bootstrap returned by kallisto or salmon pseudoaligners.
}
\examples{

       PathSamples<-system.file("extdata",package="EventPointer")
       PathSamples <- paste0(PathSamples,"/output")
       PathSamples <- dir(PathSamples,full.names = TRUE)
       
       data_exp <- getbootstrapdata(PathSamples = PathSamples,type = "kallisto")


}
