% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbootstrapkallisto.R
\name{getbootstrapkallisto}
\alias{getbootstrapkallisto}
<<<<<<< HEAD
\title{GetbootstrapKallisto}
=======
\title{getbootstrapkallisto}
>>>>>>> primersdesign
\usage{
getbootstrapkallisto(pathValues = NA, nb)
}
\arguments{
\item{pathValues}{A vector with the complete directory to the folder of the output of kallisto}

\item{nb}{number of bootstrap}
}
\value{
A list containing the quantification data and with the bootrap information.
}
\description{
Function to load the values of the bootstrap returned by kallisto pipeline

Inputs:
}
\details{
Return from the output of kallisto the data with the bootstrap information.
}
\examples{

   PathFiles <- system.file('extdata',package='EventPointer')
   PathFiles <- dir(paste0(PathFiles,'/output'),full.names = TRUE)

   #load the data

   mydatab <- getbootstrapkallisto(pathValues = PathFiles,nb = 20)


}
