% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCounts.R
\name{GetCounts}
\alias{GetCounts}
\title{Internal function called by EvenPointer}
\usage{
GetCounts(Events, sg_txiki, type = "counts")
}
\arguments{
\item{Events}{Detected alternative splicing events for a particular gene}

\item{sg_txiki}{Splicing graph of the gene}

\item{type}{Either counts of fpkms}
}
\value{
matrix with Counts/FPKMs for alternative splicing events
}
\description{
Function to obtain either counts or fpkms for each of the three paths of an alternative splicing event
}
\keyword{internal}

