% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventPaths.R
\name{getEventPaths}
\alias{getEventPaths}
\title{Internal function called by EvenPointer}
\usage{
getEventPaths(Events, SG)
}
\arguments{
\item{Events}{Detected alternative splicing events for a particular gene}

\item{SG}{Splicing graph representation for a particular gene}
}
\value{
Events list with Type slot
}
\description{
Function to get the corresponding exons and junctions for each of the paths (P1,P2,Ref) for an
alternative splicing event detected by EventPointer
}
\keyword{internal}

