% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIGVPaths.R
\name{GetIGVPaths}
\alias{GetIGVPaths}
\title{Internal function called by EvenPointer}
\usage{
GetIGVPaths(EventInfo, SG_Edges)
}
\arguments{
\item{EventInfo}{Information related to the event that is being displayed}

\item{SG_Edges}{data.frame with the edges for the splicing graph}
}
\value{
matrix with information to be passed to igv
}
\description{
Function to define the segments to be displayed in IGV to represent an alternative splicing event
}
\keyword{internal}

