% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSI.R
\name{getPSI}
\alias{getPSI}
\title{Internal function called by EvenPointer}
\usage{
getPSI(ExFit)
}
\arguments{
\item{ExFit}{data.frame with expression of alternative splicing events}
}
\value{
matrix with PSI values along events and samples
}
\description{
Function to calculate the PSI for a particular alternative splicing event
}
\keyword{internal}

