% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSI_RNASeq.R
\name{getPSI_RNASeq}
\alias{getPSI_RNASeq}
\title{Internal function called by EvenPointer}
\usage{
getPSI_RNASeq(Result)
}
\arguments{
\item{ExFit}{data.frame with fpkms of alternative splicing events}
}
\value{
matrix with PSI values along events and samples
}
\description{
Function to calculate the PSI for a particular alternative splicing event
}
\keyword{internal}

