% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomFlow.R
\name{getRandomFlow}
\alias{getRandomFlow}
\title{Internal function called by EvenPointer}
\usage{
getRandomFlow(Incidence, ncol = 1)
}
\arguments{
\item{Incidence}{Incidence matrix of the splicing graph}

\item{ncol}{number of columns for the final matrix (>1 will return a matrix of fluxes rather than a vector)}
}
\value{
matrix/vector of fluxes traversing the splicing graph
}
\description{
Function to determine the flux that is traversing each of the edges in the splicing graph
}
\keyword{internal}

