% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHsummarization.R
\name{IHsummarization}
\alias{IHsummarization}
\title{Irwin Hall Summarization for Pvalues}
\usage{
IHsummarization(Pv1, t1, Pv2, t2, coherence = "Opposite")
}
\arguments{
\item{Pv1}{Pvalue}

\item{t1}{t-statistic associated to Pv1}

\item{Pv2}{Pvalue}

\item{t2}{t-statistic associated to Pv2}
}
\value{
Pvalues list with summarized pvalues and associated z values
}
\description{
Summarization of pvalues according to the irwin-hall distribution
}
\keyword{internal}

