% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPrimers.R
\name{MyPrimers}
\alias{MyPrimers}
\title{Data frame with primers design for conventional PCR}
\format{A \code{data.frame} object displays the relative
information for primers design for conventional PCR}
\usage{
data(MyPrimers)
}
\value{
MyPrimers object contains a data.frame with
the information of the design primers for conventioanl
PCR.
}
\description{
Data frame with primers design for conventional PCR
}
