% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist2}
\alias{pdist2}
\title{Distance between matrices}
\usage{
pdist2(X, Y)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix}
}
\value{
Returns a matrix of distances
}
\description{
Calculates the distance between two matrices given by the user
}
\keyword{internal}

