% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareCountData.R
\name{PrepareCountData}
\alias{PrepareCountData}
\title{Internal function called by EvenPointer}
\usage{
PrepareCountData(Result)
}
\arguments{
\item{Result}{All alternative splicing events detected by EventPointer}
}
\value{
Matrix of counts/fpkms
}
\description{
Function to create a matrix with all the counts/fpkms per path and sample detected by EventPointer.
}
\keyword{internal}

