% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareOutput.R
\name{PrepareOutput}
\alias{PrepareOutput}
\title{Internal function called by EvenPointer}
\usage{
PrepareOutput(Result, Final)
}
\arguments{
\item{Result}{Detected alternative splicing events for a particular gene}

\item{Final}{Statistical values for the tests made to the events}
}
\value{
data.frame with results from EventPointer analysis
}
\description{
Function to create and organize the data.frame that is given to the user as output
of EventPointer function
}
\keyword{internal}

