% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareProbes.R
\name{PrepareProbes}
\alias{PrepareProbes}
\title{Internal function called by EvenPointer}
\usage{
PrepareProbes(Probes, Class)
}
\arguments{
\item{Probes}{Probes.txt file passed to EventPointer}

\item{Class}{Specify either PSR for Exon probe or Junction for junction probes}
}
\value{
Probes arranged in required format
}
\description{
Function to arrange the .txt files of exon and junction probes in the format required
by EventPointer.
}
\keyword{internal}

