% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSI_boots.R
\docType{data}
\name{PSI_boots}
\alias{PSI_boots}
\title{Bootstrap PSI values from BAM files}
\format{
A 3-dimensional array with dimensions (events, bootstrap samples, samples).
The first layer contains original PSI estimates, followed by bootstrap replicates.
}
\source{
Generated from EventsDetection_BAM function using example BAM files.
}
\usage{
PSI_boots
}
\description{
Example dataset containing PSI (Percent Spliced In) values with bootstrap
replicates obtained from RNA-Seq BAM files analysis.
}
\examples{
data(PSI_boots)
dim(PSI_boots)

}
\keyword{datasets}
