% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSIss.R
\name{PSIss}
\alias{PSIss}
\title{relationship between isoforms and events}
\description{
Relationship between isoforms and events.
}
\usage{
data(PSIss)
}
\format{
A \code{} object \code{PSIss[[1]]} displays the values of PSI and \code{PSIss[[2]]} the valeus of expression.
}
\value{
PSIss object the values of PSI calculated by the funcion \code{GetPSI_FromTranRef} and also the values of expression.
}
