% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG_Info.R
\name{SG_Info}
\alias{SG_Info}
\title{Internal function called by EvenPointer}
\usage{
SG_Info(SG_Gene)
}
\arguments{
\item{SG_Gene}{GRanges object with elements that build up the splicing graph.}
}
\value{
List with slots adjacency,incidence and Edges
}
\description{
Function to create the adjacency matrix, incidence matrix and edge relation to represent
the splicing graph of a particular gene.
}
\keyword{internal}

