% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG_reclassify.R
\name{SG_reclassify}
\alias{SG_reclassify}
\title{Splicing graph example for Events_ReClassification function}
\description{
Splicing graph example for Events_ReClassification function
}
\usage{
data(SG_reclassify)
}
\format{
A \code{list} object \code{SG_reclassify[[i]]} displays the  splicing graph of the ith gene.
}
\value{
A list with the splicing graph of the 5 genes corresponding to the alternative splicing events depicted in the example of the  function \code{Events_ReClassification}.
}
