% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteGTF.r
\name{WriteGTF}
\alias{WriteGTF}
\title{Internal function called by EvenPointer}
\usage{
WriteGTF(PATH, Data, Probes, Paths)
}
\arguments{
\item{PATH}{Directory where to write the .gtf file}

\item{Data}{Basic information of the event (Gene, Type,...,etc)}

\item{Probes}{Probes mapped to any of the paths of an alternative splicing event}

\item{Paths}{Paths that represent an alternative splicing event}
}
\value{
Adds lines to the corresponding .GTF file
}
\description{
Function to write the GTF file in the requried format to be loaded into IGV
}
\keyword{internal}

