% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteGTF_RNASeq.R
\name{WriteGTF_RNASeq}
\alias{WriteGTF_RNASeq}
\title{Internal function called by EvenPointer}
\usage{
WriteGTF_RNASeq(PATH, Data, Paths)
}
\arguments{
\item{PATH}{Directory where to write the .gtf file}

\item{Data}{Basic information of the event (Gene, Type,...,etc)}

\item{Paths}{Paths that represent an alternative splicing event}
}
\value{
Adds lines to the corresponding .GTF file
}
\description{
Function to write the GTF file in the requried format to be loaded into IGV
}
\keyword{internal}

