% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllEvents_RNASeq.R
\name{AllEvents_RNASeq}
\alias{AllEvents_RNASeq}
\title{Alternative splicing events detected by EventPointer}
\format{A \code{list} object \code{AllEvents_RNASeq[[i]][[j]]}
displays the jth splicing event for the ith gene.}
\usage{
data(AllEvents_RNASeq)
}
\value{
AllEvents_RNASeq object contains all the detected
alternativesplicing events using EventPointermethodology.
The splicing events where detected using the BAM files from
the dataset published in Seshagiri et al. 2012 andused in
the \code{SGSeq} R package vignette.
}
\description{
Alternative splicing events detected by EventPointer
}
