% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventDetection.R
\name{EventDetection}
\alias{EventDetection}
\title{Detect splicing events using EventPointer methodology}
\description{
Identification of all the alternative splicing events in the splicing graphs.
}
\usage{
EventDetection(Input, cores, Path)
}
\arguments{
\item{Input}{Output of \code{PrepareBam_EP}.}

\item{cores}{Number of cores used for parallel processing.}

\item{Path}{Directory where to write the EventsFound_RNASeq.txt file.}
}
\value{
List with all the events found for all the genes present in the experiment and a file called EventsFound_RNASeq.txt with the information of each event.
}
\examples{
\dontrun{
  # Run EventDetection function
   data(SG_RNASeq)
   TxtPath<-tempdir()
   AllEvents_RNASeq<-EventDetection(SG_RNASeq,cores=1,Path=TxtPath)
   }
}
