% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventPointerBAM_IGV.R
\name{EventPointerBAM_IGV}
\alias{EventPointerBAM_IGV}
\title{EventPointerBAM IGV Visualization}
\description{
Alternative splicing events detected by EventPointerBAM can be visualized using the \code{EventPointerBAM_IGV()} function. This function displays the paths that make up the splicing graph of each event in a gtf format that can be viewed using the IGV program. Each event is composed of 3 paths: path1, path2, and Reference, which will be indicated along with the event id for identification as “path1_”, “path2_”, “Ref_1”, and “Ref_2”.
}
\usage{
EventPointerBAM_IGV(SG_RNASeq, 
                    EventsTxt, 
                    PathGTF)
}
\arguments{

\item{SG_RNASeq}{Data.frame generated by \code{EventsDetection_BAM} with the events to be included in the GTF file.}

\item{EventsTxt}{Reference transcriptome. Must be one of: “Ensembl”, “UCSC”, “AffyGTF” or “CustomGTF”.}

\item{PathGTF}{Folder where to save the generated GTF file.}
}
\value{
Creates paths_RNASeq.gtf file that represents the alternative splicing events.
}
\examples{
\dontrun{
  EventsTxt<-paste(system.file("extdata",package="EventPointer"),"/TotalEventsFound.csv",sep="")
  load(paste(system.file("extdata",package="EventPointer"),"/SgFC.RData",sep=""))
  SG_RNASeq <- SgFC
  PathGTF <- tempdir()
  
  EventPointerBAM_IGV(SG_RNASeq, 
                      EventsTxt, 
                      PathGTF)
}
}