% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_EP_mod.R
\name{EventPointerStats_BAM}
\alias{EventPointerStats_BAM}
\title{This function performs the differential comparison between sample groups for the calculation of delta PSI from EventDetection_BAM function}
\usage{
EventPointerStats_BAM(PSI_boots, Design, Contrast, cores=8, UseBootstrap=F,
                  Threshold = 0, nbootstraps = 1000, pathResult = PathSGResult)
}
\arguments{

\item{Design}{A matrix defining the linear model. Each row corresponds to an array, and each column corresponds to a coefficient (such as the baseline and treatment effects).}

\item{Contrast}{A numeric matrix with contrasts to be tested. Rows correspond to coefficients in the design matrix, and columns correspond to contrasts.}

\item{Threshold} {it assigns a threshold to compute the pvalues. Default = 0}

\item{UseBootstrap}{ Boolean variable to indicate if bootstrap step in PSI is applicated.}

\item{nbootstraps}{How many layers, Bootstraps or samplings are goingto be used. Caution, high numbers increase computational time.}

\item{cores}{Number of cores used for parallel processing}

\item{pathResult}{Folder where save the function result.}
}
\value{
Folder containing .csv files indicating the events and differential statistics for each contrast specified in the contrast matrix.
}
\description{

}
\examples{
\dontrun{
  # Run EventDetection function
   data(PSI_boots)
   Design <- cbind(rep(1,9),
                 rep(c(1,0,0),3),
                 rep(c(0,1,0),3))
  Contrast <- cbind(
    c(0,1,0),
    c(0,0,1))
   PathSGResult<-tempdir()
   EventPointerStats_BAM(PSI_boots, Design, Contrast, cores=8, UseBootstrap=F,
                  Threshold = 0, nbootstraps = 1000, pathResult = PathSGResult)

}
