% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_EP_mod.R
\name{EventPointerStats_ST}
\alias{EventPointerStats_ST}
\title{This function performs the differential comparison between sample groups for the calculation of delta PSI from EventPointerStats_ST function}
\usage{
EventPointerStats_ST(PSI, Design, Contrast, cores=1, 
                                   BootstrapStats = T,nbootstraps= 10000, 
                                   UsePseudoAligBootstrap = T,Threshold = 0,
                                   pathResult=pathResult)
}
\arguments{

\item{PSI}{RData variable resulting  from EventsDetection_ST()}

\item{Design}{A matrix defining the linear model. Each row corresponds to an array, and each column corresponds to a coefficient (such as the baseline and treatment effects).}

\item{Contrast}{A numeric matrix with contrasts to be tested. Rows correspond to coefficients in the design matrix, and columns correspond to contrasts.}

\item{Threshold} {it assigns a threshold to compute the pvalues. Default = 0}

\item{UsePseudoAligBootstrap}{ TRUE (default) if bootstrap data from pseudoaligment want to be used or FALSe if not.}

\item{nbootstraps}{How many layers, Bootstraps or samplings are goingto be used. Caution, high numbers increase computational time.}

\item{cores}{Number of cores used for parallel processing}

\item{pathResult}{Folder where save the function result.}
}
\value{
Folder containing .csv files indicating the events and differential statistics for each contrast specified in the contrast matrix.
}
\description{

}
\examples{
\dontrun{
  data(PSI_boots)
   
  #Creation of design matrix
  Design <- cbind(1,rep(c(0,1),each=2))
  #Creation of contrast matrix
  Contrast <- matrix(c(0,1),nrow=2)
  
  pathResult <- tempdir()
  EventPointerStats_ST(PSI, Design, Contrast, cores=1, 
                                   BootstrapStats = T,nbootstraps= 10000, 
                                   UsePseudoAligBootstrap = T,Threshold = 0,
                                   pathResult=pathResult)

}
