% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_EP_mod.R
\name{EventPointerStats_ST}
\alias{EventPointerStats_ST}
\title{Differential comparison between sample groups for the calculation of delta PSI}
\usage{
EventsPSI <- EventsDetection_ST(PathSamplesAbundance, 
                                typeAbundance = "kallisto",
                                PathTranscriptomeGTF = PathTranscriptomeGTF, 
                                EventsTranscriptome=NULL, 
                                Bootstrap=T, 
                                Filter=F, 
                                Qn = 0.25, 
                                cores=1,
                                PathEventsGTFResults=Pathtxt)
}
\description{
This function performs the differential comparison between sample groups for the calculation of delta PSI from \code{EventDetection_ST} function.
}

\arguments{

\item{PSI}{PSI_ST.RData variable resulting from \code{EventsDetection_ST()}.}

\item{Design}{A matrix defining the linear model. Each row corresponds to an array, and each column corresponds to a coefficient (such as the baseline and treatment effects).}

\item{Contrast}{A numeric matrix with contrasts to be tested. Rows correspond to coefficients in the design matrix, and columns correspond to contrasts.}

\item{BootstrapStats}{Boolean variable to indicate if bootstrap step in PSI is applied.}

\item{nbootstraps}{How many layers, Bootstraps or samplings are goingto be used. Caution, high numbers increase computational time.}

\item{UsePseudoAligBootstrap}{TRUE (default) if bootstrap data from pseudoaligment want to be used or FALSe if not.}

\item{Threshold}{It assigns a threshold to compute the pvalues. Default value is 0.}

\item{cores}{Number of cores used for parallel processing.}

\item{pathResult}{Folder where save the function result.}
}
\value{
Folder containing .csv files with the results associated with the differential \eqn{\Psi} analysis of the indicated contrasts.It presents the \eqn{\Delta \Psi} associated with each event of that contrast and its corresponding significance parameters. Note that a table will be obtained for each contrast indicated in the contrast matrix.

}

\examples{
\dontrun{
  # Run EventsDetection_ST function
  data(PSI_boots)
  
  Design <- cbind(1,rep(c(0,1),each=2))
  Contrast <- matrix(c(0,1),nrow=2)
  
  pathResult <- tempdir()
  
  EventPointerStats_ST(PSI, 
                     Design, 
                     Contrast, 
                     BootstrapStats = T,
                     nbootstraps= 10000,
                     UsePseudoAligBootstrap = T,
                     Threshold = 0,
                     cores=1, 
                     ram = 4, 
                     pathResult=pathResult)
}
}