% Generated by roxygen2: do not edit by hand
\name{EventsDetection_ST}
\alias{EventsDetection_ST}
\title{Detect splicing events using EventsDetection_ST workflow}
\usage{
EventsDetection_ST(PathSamplesAbundance = NULL, PathTranscriptomeGTF = NULL, EventsTranscriptome = NULL,
                              PathEventsGTFResults=".",
                              cores=1, typeAbundance = "kallisto", Bootstrap=T, Filter=F,
                              Qn = 0.25)
}
\arguments{

\item{PathSamplesAbundance}{Path to the folder with the alignment .bam files. It can be the path to a folder that contains other folders with the bams.}

\item{PathTranscriptomeGTF}{Path to the GTF file of the reference transcriptome.}

\item{EventsTranscriptome}{If the user wishes to provide events detected from a previous work, the path to the RData file can be provided.}

\item{PathEventsGTFResults}{Path to locate the results of the function.}

\item{cores}{Number of cores used for parallel processing}

\item{typeAbundance}{Type of abundance files provided for the PSI analysis. The options are "salmon" or "kallisto".}
}
\item{Bootstrap}{	Boolean variable to indicate if bootstrap data from pseudo-alignment is used.}

\item{Filter}{	Boolean variable to indicate if an expression filter is applied. Default FALSE.}

\item{Qn}{Quartile used to filter the events (Bounded between 0-1, Qn would be 0.25 by default).}

\value{
List with the information of all events found in samples.
}
\description{

}
\examples{
\dontrun{
  # Run EventDetection function
  PathFiles<-system.file("extdata",package="EventPointer")
  PathTranscriptomeGTF <- paste(PathFiles,"/gencode.v24.ann_2genes.gtf",sep="")
  PathSamplesAbundance <- paste0(PathFiles,"/output")
  PathSamplesAbundance <- dir(PathSamplesAbundance,full.names = TRUE)
  Pathtxt <- tempdir()
  EventsPSI <- EventsDetection_ST(PathSamplesAbundance,PathTranscriptomeGTF = PathTranscriptomeGTF,
                              PathEventsGTFResults=Pathtxt,
                              cores=1, typeAbundance = "kallisto", Bootstrap=T, Filter=F,
                              Qn = 0.25)

}
