% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventXtrans.R
\name{EventXtrans}
\alias{EventXtrans}
\title{relationship between isoforms and events}
\description{
relationship between isoforms and events
}
\usage{
data(EventXtrans)
}
\format{
A \code{list} object \code{EventXtrans[[1]]} displays the isoform that build up the path1 of each event.
}
\value{
EventXtrans object contains the relationship between the isoforms and the events. It is a list of 4 elements. the first three stored sparse matrices relating the isoforms with the events. The fourth element stores de names of the reference annotation used (isoforms names).
}
