% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Fit}
\alias{Fit}
\title{Result of EventPointer_Bootstrap}
\format{
A \code{list} object
}
\usage{
data(Fit)
}
\value{
A list containing the summary of the Bootstrap analysis: DeltaPSI, Pvalues, FDR. This info can be
obtained in a simple table with the function ResulTable.
}
\description{
Result of EventPointer_Bootstrap
}
