% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{annotateEvents}
\alias{AnnotateEvents_RNASeq}
\alias{ClassifyEvents}
\alias{estimateAbsoluteConc}
\alias{findTriplets}
\alias{GetCounts}
\alias{getEventPaths}
\alias{GetIGVPaths}
\alias{getPSI}
\alias{getPSI_RNASeq}
\alias{getRandomFlow}
\alias{IHsummarization}
\alias{pdist2}
\alias{PrepareCountData}
\alias{PrepareProbes}
\alias{PrepareOutput}
\alias{SG_Info}
\alias{WriteGTF}
\alias{WriteGTF_RNASeq}
\alias{flat2Cdf}
\title{EventPointer Internal Functions}
\usage{
annotateEvents(Events, PSR_Gene, Junc_Gene, Gxx)

AnnotateEvents_RNASeq(Events)

ClassifyEvents(SG, Events)

estimateAbsoluteConc(Signal1, Signal2, SignalR, lambda)

findTriplets(randSol, tol = 1e-08)

GetCounts(Events, sg_txiki, type = "counts")

getEventPaths(Events, SG)

GetIGVPaths(EventInfo, SG_Edges)

getPSI(ExFit)

getPSI_RNASeq(Result)

getRandomFlow(Incidence, ncol = 1)

IHsummarization(Pv1, t1, Pv2, t2, coherence = "Opposite")

pdist2(X, Y)

PrepareCountData(Result)

PrepareProbes(Probes, Class)

PrepareOutput(Result, Final)

SG_Info(SG_Gene)

WriteGTF(PATH, Data, Probes, Paths)

WriteGTF_RNASeq(PATH, Data, Paths)

flat2Cdf(file, chipType, tags = NULL, rows = 2560, cols = 2560,
  verbose = 10, xynames = c("X", "Y"), gcol = 5, ucol = 6, splitn = 4,
  col.class = c("integer", "character")[c(1, 1, 1, 2, 2, 2)], ...)
}
\value{
Internal outputs
}
\description{
Internal functions used by EventPointer in the different steps of the algorithm
}
\keyword{internal}
