% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{AnnotateEvents_RNASeq}
\alias{ClassifyEvents}
\alias{GetCounts}
\alias{GetIGVPaths}
\alias{IHsummarization}
\alias{InternalFunctions}
\alias{PrepareCountData}
\alias{PrepareOutput}
\alias{PrepareProbes}
\alias{SG_Info}
\alias{WriteGTF}
\alias{WriteGTF_RNASeq}
\alias{annotateEvents}
\alias{estimateAbsoluteConc}
\alias{findTriplets}
\alias{getEventPaths}
\alias{getPSI}
\alias{getPSI_RNASeq}
\alias{getRandomFlow}
\alias{pdist2}
\title{EventPointer Internal Functions}
\usage{
annotateEvents(Events, PSR_Gene, Junc_Gene, Gxx)

AnnotateEvents_RNASeq(Events)

ClassifyEvents(SG, Events)

estimateAbsoluteConc(Signal1, Signal2, SignalR, lambda)

findTriplets(randSol, tol = 1e-08)

GetCounts(Events, sg_txiki, type = "counts")

getEventPaths(Events, SG)

GetIGVPaths(EventInfo, SG_Edges)

getPSI(ExFit)

getPSI_RNASeq(Result)

getRandomFlow(Incidence, ncol = 1)

IHsummarization(Pv1, t1, Pv2, t2, coherence = "Opposite")

pdist2(X, Y)

PrepareCountData(Result)

PrepareProbes(Probes, Class)

PrepareOutput(Result, Final)

SG_Info(SG_Gene)

WriteGTF(PATH, Data, Probes, Paths)

WriteGTF_RNASeq(PATH, Data, Paths)
}
\value{
Internal outputs
}
\description{
Internal functions used by EventPointer in the different steps of the algorithm
}
\keyword{internal}

