% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSIss.R
\name{PSIss}
\alias{PSIss}
\title{relationship between isoforms and events}
\format{
A \code{} object \code{PSIss[[1]]} displays the values of PSI
and \code{PSIss[[2]]} the valeus of expression.
}
\usage{
data(PSIss)
}
\value{
PSIss object the values of PSI calculated by the funcion
GetPSI_FromTranRef and also the values of expression.
}
\description{
relationship between isoforms and events
}
