% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxD.R
\name{TxD}
\alias{TxD}
\title{Transcript x Protein Domain matrix: small matrix for examples}
\format{
A \code{matrix} object
}
\usage{
data(TxD)
}
\value{
A matrix containing the relates Transcripts with Protein Domains
}
\description{
Transcript x Protein Domain matrix: small matrix for examples
}
